rcll: Keep track of successfully PROCESSED production steps
Fact template (product-processed) tracks a single
successfully PROCESSED step  done at an mps. Besides
the info about the step, it tracks the ID of the workpiece
that was available at the  machine. and  weather the step
was CONFIRMED.

The fact will be used as the central entity to control
awarding score for an intermediate production steps.

Intended flow:
Initially, the product-processed NOT CONFIRMED, meaning
the fact is not linked to any workpiece. When workpiece
tracking is enabled, it will be confirmed upon
AVAILABILITY of a workpiece at the expected machine. When
workpiece tracking is disabled, the workpiece will be
simulated on delivery creating all confirmed steps that
contributed into the production of the workpiece.