rcll: deffunc to simulate workpiece tracking
In order to, award the intermediate production points
on delivery, simulate all product-processing steps and
relate it to a fresh workpiece.

(for example, if workpiece tracking is not enabled, we still
want to be awarded the points given for intermediate steps,
upon delivery of the product. Which is the case atm since
tracking is not fully implemented and not yet linked to barcode
reader)