rcll: Extend assign/resign an orders to a workpiece
A workpiece is assigned to any order that matches a perfect
prefix of it colors. A workpiece is resigned from an order
whenever it becomes inconsistent with it.

The Flow of assigning an order to WP:
This means that the workpiece will be assigned a 'best guess'
order directly after it gets its base (ie, 'any' order that
has the same base). When a ring is mounted, the order will be
one step closer to the right guess and the old order will be
resigned as it becomes inconsistent with the recent operation.
(For Example, when starting to work on a C2 product with a red
base, at first the workpiece could be assigned to some C0 with
the same base color. As the workpiece gets its first ring, the
C0 order will be resigned off the workpiece and the correct C2
order will be assigned [see,assumption 1]. If the team then
mounts a wrong second ring, the WP will be resigned from the
C2 order.

This assignment is correct based on the following assumptions
and grantees.

[assumption 1]
 All Cx complexities have a unique ring color prefix. ie, the
 color of the first ring immediately determine the order of the
 workpiece.

[assumption 2]
 There is no harm in assigning the non intended 'valid' order
 temporarily as long as the mistake is rectified as soon as it
 becomes clear.
 -For a Cx that happens by mounting the first ring.
 -For C0's with a different cap that happens when mounting the
  cap.
 -For 2 identical products that happens when the delivery is
  requested.

[assumption 3]
 Any valid ambiguity will be corrected latest on the delivery
 station when the bot requests the delivery of the order.

[assumption 4]
 The referee confirmation is ground truth. When the referee
 confirms the order of the workpiece, any 'wrong' assignment
 will  and rectified and any intermediate points given for the
 'wrong' order will be replaced by the those of the correct order.

Fact:
 The case of 2 identical orders (could happen with C0 for example).
 The only way to 'currently' know which one was intended is upon
 delivery.