rcll: Link the delivery step at DS to an available workpiece
When a delivery is attempted, the order-id of the tracked
WP could potentially be inconsistent with the order requested
for delivery at DS. This could either be due to a wrong delivery
by the team, or a tracking failure!

This will be Verified later on, when the referee confirms the
order and what the workpiece looks like

ie,This product-processed step will be CONFIRMED later on
(by referee decision).