rcll: Simulating a workpiece does not create the delivery step
When simulating workpiece tracking, assert all production
steps except the product-processed fact of done the DS
stations (aka, the delivery)

Always assert the product-processed step of the delivery
(even when WP-tracking is disabled. This will be done in
the next commits.
The idea will be to always track the delivery step much
like what (product-delivered) fact currently does. The
product-delivered fact will be refactored in the next
commits