rcll: rm steps and points given for a deleted workpiece
If a workpiece is deleted, remove all the product-processed
steps related to it and any points given. Will only happen
after a referee confirms the WP to be different than what
we believe it is.

Only keep the  product-processed at DS of the delivery since
it is still valid. It will be related to the fresh simulated WP